requirements.txt
# Titan Terminal Backend Dependencies

# API Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0

# Anthropic Claude
anthropic>=0.18.0

# HTTP Client
httpx>=0.26.0

# Environment
python-dotenv>=1.0.0

# Database
# SQLite is built into Python

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
